+++ packages/dara-core/js/actions/update-variable.tsx
@@ -1,5 +1,5 @@
-import { getAtom } from '@/shared/interactivity';
-import { resolveNested, setNested } from '@/shared/interactivity/nested';
+import { getOrRegisterPlainVariable } from '@/shared/interactivity/plain-variable';
+import { getOrRegisterUrlVariable } from '@/shared/interactivity/url-variable';
 import { ActionHandler, UpdateVariableImpl } from '@/types/core';
 
 /**
@@ -12,55 +12,39 @@ export const INPUT = '__dara_input__';
  */
 export const TOGGLE = '__dara_toggle__';
 
-function isFunction(val: any): val is (args: any) => any {
-    return typeof val === 'function';
-}
-
-/**
- * Get an updater function for the variable.
- * For plain variables with a nested property this will return a function that will update the nested property.
- *
- * @param variable variable to update
- * @param updater value or function to update the variable with
- */
-function getUpdater<T = any>(
-    variable: UpdateVariableImpl['variable'],
-    updater: T | ((value: T) => T)
-): T | ((oldValue: T) => T) {
-    // is a Variable().get()
-    if (variable.__typename === 'Variable' && variable.nested && variable.nested.length > 0) {
-        const nested = variable.nested.map((n) => String(n));
-        return (oldValue: any) => {
-            let newValue = updater;
-            if (isFunction(updater)) {
-                newValue = updater(resolveNested(oldValue, nested));
-            }
-            return setNested(oldValue, nested, newValue);
-        };
-    }
-
-    // return as-is
-    return updater;
-}
-
 /**
  * Front-end handler for UpdateVariable action.
  */
 const UpdateVariable: ActionHandler<UpdateVariableImpl> = (ctx, actionImpl) => {
-    const varAtom = getAtom(actionImpl.variable);
+    let varAtom;
+
+    // Make sure the variable is registered
+    switch (actionImpl.variable.__typename) {
+        case 'Variable':
+            varAtom = getOrRegisterPlainVariable(
+                actionImpl.variable,
+                ctx.wsClient,
+                ctx.taskCtx,
+                ctx.location.search,
+                ctx.sessionToken
+            );
+            break;
+        case 'UrlVariable':
+            varAtom = getOrRegisterUrlVariable(actionImpl.variable);
+            break;
+        case 'DataVariable':
+            throw new Error('DataVariable is not supported in UpdateVariable action');
+    }
 
     if (actionImpl.value === INPUT) {
-        return ctx.set(varAtom, getUpdater(actionImpl.variable, ctx.input));
+        return ctx.set(varAtom, ctx.input);
     }
 
     if (actionImpl.value === TOGGLE) {
-        return ctx.set(
-            varAtom,
-            getUpdater(actionImpl.variable, (value) => !value)
-        );
+        return ctx.set(varAtom, (value) => !value);
     }
 
-    return ctx.set(varAtom, getUpdater(actionImpl.variable, actionImpl.value));
+    return ctx.set(varAtom, actionImpl.value);
 };
 
 export default UpdateVariable;